examples/prompt-based-agent/requirements.txt
fastapi[standard]
fastmcp
pydantic
googlesearch-python
json-repair
openai
playwright